requirements.txt
# Core HTTP + environment
requests
python-dotenv

# Hugging Face summarisation stack (BART large CNN)
transformers
torch
sentencepiece

# Posting client for X/Twitter
tweety-ns
tweepy
